refactor(steam): unify sign-in result handling and consolidate app id
