--- apps/stage-tamagotchi/src/main/services/steam/app-id.ts
@@ -1,1 +1,0 @@
-export const STEAM_APP_ID = 3885340
