+++ apps/stage-tamagotchi/src/main/services/steam/types.ts
@@ -1,3 +1,5 @@
+export const STEAM_APP_ID = 3885340
+
 export type SteamInitResult
   = | { ok: true }
     | { ok: false, reason: 'not_steam' | 'init_failed' | 'api_unavailable' }
